add modules and setup the database fields

+++ schedule/scripts/fake_data.py
@@ -14,10 +14,11 @@ if __name__ == '__main__':
     os.environ.setdefault("DJANGO_SETTINGS_MODULE", "schedule.settings")
     django.setup()
     from account.models import CustomUser as User
-    print('clean database')
-    User.objects.all().delete()
+    if User.objects.all():
+        print('clean database')
+        User.objects.all().delete()
 # superuser
-    print('create a super user')
-    user = User.objects.create_superuser(
+        print('create a super user')
+        user = User.objects.create_superuser(
         'circlepen', 'lyle.lai@redfalcon-hpc.com', 'redfalcon')
     print('finish')
